+++ src/scripts/index.js
@@ -1,27 +1,8 @@
-export {cardButton,imageContainer,editUserInfoButton,placesCardList,popupImage,valid,userNameInfo,userJobInfo,api,group};
 import '../pages/index.css';
-import Card from './card.js';
-import Popup from './popUp.js';
+import './script.js';
 import Validation from './valid.js';
-import {auth} from './api.js';
-import Api from './api.js';
-
-  const userNameInfo = document.querySelector('.user-info__name');
-  const userJobInfo = document.querySelector('.user-info__job');
-  const api = new Api();
-  const group = 'cohort3';
-  const cardButton = document.querySelector('.popup__button');
-  const openForm = document.querySelector('.user-info__button');
-  const formContainer = document.querySelector('.popup');
-  const closeForm = document.querySelector('.popup__close');
-  const openEditForm = document.querySelector('.button-edit');
-  const editFormContainer = document.querySelector('.popup-edit');
-  const placesCardList = document.querySelector('.places-list');
-  const editUserInfoButton = editFormContainer.querySelector('.popup__button-edit');
-  const closeEditForm = editFormContainer.querySelector('.popup__close');
-  const imageContainer = document.querySelector('.popup-image');
-  const popupImage = document.querySelector('.popup__image');
-  const closeImageForm = imageContainer.querySelector('.popup__close');
+import Card from './card.js';
+  
   const popupInput = document.querySelector('.popup__input_type_name');
   const popupInputLink = document.querySelector('.popup__input_type_link-url');
   const editButton = document.querySelector('.popup__button-edit');
@@ -35,23 +16,9 @@ import Api from './api.js';
   const formInputLink = formNew.elements.link;
   const formInputEditName = formEdit.elements.userName;
   const formInputInfo = formEdit.elements.info;
-  const popupForm = new Popup(formContainer, closeForm, openForm);
-  const popupFormAccept = new Popup(formContainer, cardButton, openForm);
-  const popupEdit = new Popup(editFormContainer, closeEditForm, openEditForm);
-  const popupEditAccept = new Popup(editFormContainer, editUserInfoButton, openEditForm);
-  const popupImg = new Popup(imageContainer, closeImageForm, popupImage);
-  const validation = new Validation();
-  const valid = {
-    validationLenght: 'Должно быть от 2 до 30 символов',
-    notEmpty: 'Поле не должно быть пустым'
-  }
-  api.getUser(group, auth);
-  api.getArray(group, auth);
+  const cardButton = document.querySelector('.popup__button');
+  const validation = new Validation(); 
   
-
-  
-  
-    
   
   //создание карточки 
   cardButton.addEventListener('click', function () {
@@ -72,7 +39,7 @@ import Api from './api.js';
   infoInput.addEventListener('input', function () {
     validation.checkValidity(formInputInfo, errorInfo)
   });
-  
+
   formNew.addEventListener('input', function () {
     validation.checkMinLength(formInputName, formInputLink, cardButton);
   });
