+++ src/scripts/script.js
@@ -1,0 +1,8 @@
+import './card.js';
+import './popUp.js';
+import './api.js';
+import './const.js';
+import './script.js';
+import './valid.js';
+
+
